+++ src/components/shared/Fields/Textarea/index.ts
@@ -1,5 +1,3 @@
-export { default } from './Textarea';
-export { Props as TextareaComponentProps } from './Textarea';
 export {
   default as HookFormTextArea,
   TextareaComponentProps as HookFormTextareaComponentProps,
